+++ src/components/custom-field-input-number/custom-field-input-number.jsx
@@ -42,6 +42,7 @@ export default function CustomFieldInputNumber(props) {
   return (
     <CustomFieldInputText
       className={getRootClassName(props.className, invalid, props.disabled)}
+      defaultValue={props.value}
       disabled={props.disabled}
       error={invalid}
       helpText={props.helpText}
@@ -58,7 +59,6 @@ export default function CustomFieldInputNumber(props) {
       required={props.required}
       step={props.step}
       type="number"
-      value={props.value}
     />
   );
 }